Add test for vault GCE auth
